# specs/001-vpn-manager-service/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: WireGuard VPN Manager API
  description: |
    云环境WireGuard VPN管理服务API
    
    ## 认证方式
    
    - **VM端API**: 使用固化Token + 来源IP双重校验
    - **外部前端应用API**: 使用Bearer Token认证
    - **管理前端API**: 使用Bearer Token认证
    
    ## 状态码
    
    - 200: 成功
    - 201: 创建成功
    - 400: 请求参数错误
    - 401: 未认证
    - 403: 权限不足
    - 404: 资源不存在
    - 409: 资源冲突
    - 500: 服务器内部错误
    
  version: 1.0.0
  contact:
    name: FCloudVPN Team
servers:
  - url: https://api.vpn-manager.internal
    description: 内网API服务器
...
paths:
...
  /api/vm/config:
    get:
      tags:
        - VM API
      summary: 获取VPN配置
      description: |
        VM首次启动时调用此接口获取WireGuard Server配置。
        
        - 系统从TCP报文头RemoteAddr提取请求来源IP
        - 如果IP不在资源池中，返回错误
        - 如果配置已存在，返回相同配置（幂等性）
        - 如果配置状态为started，拒绝请求
      security:
        - VMToken: []
      responses:
        '200':
          description: 成功返回Server端配置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerConfig'
              example:
                vm_ip: "192.168.1.100"
                vpn_ip: "10.1.100.254"
                private_key: "aBcDeFgHiJkLmNoPqRsTuVwXyZ1234567890ABCDEF="
                public_key: "xYz1234567890ABCDEFaBcDeFgHiJkLmNoPqRsT="
                listen_port: 2588
                peers:
                  - public_key: "client1PublicKey1234567890ABCDEF="
                    allowed_ips: "10.1.100.1/32"
        '400':
          description: 请求错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ip_not_in_pool:
                  value:
                    error: "该IP未在资源池中配置"
                already_started:
                  value:
                    error: "配置已启动，无法重新获取"
        '401':
          description: Token无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/frontend/configs/{vm_ip}:
    get:
      tags:
        - Frontend App API
      summary: 查询客户端配置
      description: |
        外部前端应用查询指定VM的客户端配置，展示给用户下载。
        
        返回5个客户端配置，每个配置包含备注名。
      security:
        - BearerAuth: []
      parameters:
        - name: vm_ip
          in: path
          required: true
          schema:
            type: string
            format: ipv4
          description: 虚拟机IP地址
          example: "192.168.1.100"
      responses:
        '200':
          description: 成功返回客户端配置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientConfigs'
        '404':
          description: 配置不存在或已销毁
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/frontend/configs/{vm_ip}/destroy:
    post:
      tags:
        - Frontend App API
      summary: 销毁配置
      description: |
        外部前端应用删除虚拟机时调用此接口销毁VPN配置。
        
        - 支持销毁状态1和状态2的配置
        - 销毁后配置移入归档表
      security:
        - BearerAuth: []
      parameters:
        - name: vm_ip
          in: path
          required: true
          schema:
            type: string
            format: ipv4
          description: 虚拟机IP地址
      responses:
        '200':
          description: 销毁成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "配置已销毁"
        '404':
          description: 配置不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== 管理前端API ====================
...
  /api/auth/login:
    post:
      tags:
        - Auth API
      summary: 管理员登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: "root"
                password:
                  type: string
                  format: password
                  example: "admin123"
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: "bearer"
                  expires_in:
                    type: integer
                    example: 3600
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/logout:
    post:
      tags:
        - Auth API
      summary: 管理员登出
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 登出成功
  /api/auth/me:
    get:
      tags:
        - Auth API
      summary: 获取当前用户信息
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功返回用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
components:
  securitySchemes:
    VMToken:
      type: apiKey
      in: header
      name: X-VM-Token
      description: VM端API固化Token
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer Token认证
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: 错误信息
...
    ServerConfig:
      type: object
      properties:
        vm_ip:
          type: string
          format: ipv4
        vpn_ip:
          type: string
          format: ipv4
          description: VPN网段内Server的IP
        private_key:
          type: string
          description: Server端私钥
        public_key:
          type: string
          description: Server端公钥
        listen_port:
          type: integer
          description: WireGuard监听端口
        peers:
          type: array
          items:
            type: object
            properties:
              public_key:
                type: string
              allowed_ips:
                type: string
    ClientConfigs:
      type: object
      properties:
        vm_ip:
          type: string
          format: ipv4
        status:
          type: string
          enum: [init, started]
        clients:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "client1"
              vpn_ip:
                type: string
                example: "10.1.100.1"
              public_key:
                type: string
              config_file:
                type: string
                description: 完整的WireGuard配置文件内容
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [root, admin]
        created_at:
          type: string
          format: date-time
...